added status effect!!! added ice peashooter and added methods for image library

+++ src/model/Entity.java
@@ -34,7 +34,7 @@ public abstract class Entity {
         return y;
     }
 
-    public int getSpeed(){
+    public double getSpeed(){
         return speed;
     }
